+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0 (IN PROGRESS)
 * [STHUB-1](https://folio-org.atlassian.net/browse/STHUB-1) Added config and logic for session validation. If session is invalid, redirect to keycloak login page.
+* [STHUB-3](https://folio-org.atlassian.net/browse/STHUB-3) Make an entitlement request and cache the response in local storage.
 
 ## 1.0.0